feat: isolate lab files from each other
With this change the server can run multiple
labs simultaneously. Previously labs were
written in a single directory that was
mapped into the docker container. That
included that labs could overwrite each
other easily. We could have worked around
that by using unique directory names.
However, the follwing technique is considered
superior for now.

With this approach no files are written on
the host at all. Rather we generate a
big weird looking echo command to create
the files directly inside the container.

The echo commands are concatenated with &
rather than && so that files are written
asychronously rather than sequentially.

Fixes #4