chore(dependencies): use @angular/flex-layout npm package
This commit makes sure that @angular/flex-layout is now properly fetched
from npm. Previously we had to manually clone and build the package because
it wasn't published on npm yet. This has changed since the official beta
release, thanks to @ThomasBurleson. To get this package, simply remove the
`flex-layout` folder from `node_modules/@angular` and run `(npm|yarn) install`.

There was also a *breaking change* in the package that directive properties
and selectors now follow the official Angular styleguide and are therefore
camelcased instead of dash-cased. This commit applies the changes to the
project's layout.