feat: toggle between run and stop button
With this change the run button changes
to a stop button while the lab is executing.

The stop button doesn't actually stop
the lab yet which is planned to land in a
future commit.

This change also introduces the model of
a LabExecutionContext and moves the state
(Pristine, Running etc) there. A lab may
be viewed by multiple people and while it
is running for one person it may be stopped
for another. Hence it's not really the
lab that should have these states but rather
the execution context.

Also note that the ref_id was moved to the
context which did not belong to any model
previously. Increases cohesion.

HEADS UP: This change requires the app to be
the latest performed against the latest server.

Relates to #12