feat: use firebase to stream output
With this change we are not relying on socket.io
for the communication to the client any longer.

Instead we are using firebase to listen for
scheduled runs, propagate messages through outputs
and stop executions based on changed runs that
were set to `type: 'stopped'`.

Also the tear down code does not rely on the
lab id any longer. Previously it was impossible
to have multiple runs of the same lab simultanously.

The tear down code is also much more reliable and
uses less hacks.

Closes #7, #15