feat(EditorView): add button for creating new labs
This commit adds a button to EditorViewComponent's UI which lets the user
create a new lab by navigating to '/?tpl=blank'.

To make this work, LabResolver had to be made a bit smart to figure out
how a lab needs to be created. With this commit the following rules apply:

- if a lab id is given, fetch it, if it fails, create a new one

- if no lab id is given, check if a template id is given. If given, create
  lab from template, unless template id is specified as 'blank'. Then create
  new blank one

- if no lab id and no template id is given, create lab from default template
  (which is currently the XOR sample)

It also introduces a new `NavigationConfirmDialogComponent` which is used to
make sure a new lab is only created if a user is really sure to do so.

Closes #33