feat(EditorView): add support for linking anonymous users with permanent users
As discussed in #38, anonymous users should be able to preserve their created labs
when they authenticate via GitHub. This commit introduces a new API
`AuthService.linkOrSignInWithGitHub(): Observable<User>`, which always tries
to link the anonymous user on sign in first and falls back to normal sign in if
it throws (e.g. when credentials have been linked already).

Closes #38