feat(ux): make active file configurable via URL
This commits enables users to use the `?file` query params, to configure
which file is active when the lab is loaded.

In addition, ML will now automatically update the URL with the latest active
file whenever a file is selected. That's why, as of this commit, when a user
navigates to `/SJBJxfS3g`, ML will immediately update the URL to
`/SJBJxfS3g?file=main.py` unless there's a `file` query param given.