feat(Toolbar): show lab owner info
This commit now displays the creator of a lab in the toolbar. We determine the
user of a lab by using `UserService.getUser(id)`. Anonymous users aren't stored
in our firebase, which means in case of an anonymous user, it'll emit `null`.

That's why we can rely "blindly" on Angular's `*ngIf ... else ... ;` syntax.

Once #66 landed, the template should link to the user profile.

Closes #92