refactor(Toolbar): reintroduce more generic ToolbarComponent
After renaming our previous `ToolbarComponent` to `EditorToolbarComponent`,
this commit introduces a new `ToolbarComponent` that can be used in other
application views such as the user profile view.

`ToolbarComponent` doesn't have lab actions and it also doesn't render
any lab information. It's purpose (for now) is to simply show the login avatar
and render a menu to login and logout.

Since we need the same login/logout functionality in `EditorToolbarComponent`
as well, this commit also extracts the menu + login functionality into its
own `ToolbarMenuComponent`, so it can be used in different toolbar compositions.

To not duplicate CSS, we also introduce a new `ToolbarLogoComponent` which comes
right out-of-the-box with `ToolbarComponent` and `EditorComponent`.

To make reusability across different `NgModule` easier and to support lazy-loading,
All of `Toolbar*Component`s have been extracted into their own `NgModule`,
`ToolbarModule`.

This can now be used in #86

In the near future, we might want take it one level further and make `ToolbarComponent`
aware of content projection. That allows us to remove duplicate CSS for `ml-toolbar`,
which is also introduced with this commit.