fix: don't save twice when forking
This refactors the code with the intend to
fix two bugs.

First bug:

If we fork, we directly save and then open
a edit dialog which may perform another save.

Since we allways show this dialog we can
delay the save operation until the user
is past the dialog

Second bug:

App crashes when user cancels edit action

This change makes the `EditLabDialog` dumb
in the sense that it does not perform the
actual save action anymore. Instead it will
provide enough information to the outside
world to detect if changes need to be saved
or not.

It then breaks up the code in the editor
to use the dialog in a more fine grained way.