tests(doneWhen): adds helper utility
Often in tests we have to wait until N
async operations are done before we can
make our assertions. There doesn't seem
to be a built in way to do this.

With DoneWhen one can simply wrap
the done callback, setup expectations
on when to call through and then
simply call `call` on doneWhen
as many times as needed.