feat: show rejection dialog when anonymous users try to run lab
As discussed in #110, anonymous users shouldn't be able to run labs that aren't
cached. This is already implemented, but so far the only thing we've shown was
a little snackbar saying that the execution has been rejected.

This commit introduces a new dialog component that encourages users to
request a beta invite or bake our project on Kickstarter.

Closes #110