fix: don't use set to update invocation
With our reworked firebase rules it is important
to replace `set` here with `update` to only
include fields that the user is allowed to
write to. This wasn't a problem before
because technically users could have
assigned the server themselve which clearly
should not be possible. The new firebase
security rules prevent this.

But even if we don't take the security rules into
account it is better to use `update` instead of
`set` and Object.assign(..) here.