fix: make ToolbarMenuComponent observe user auth changes
This change is mainly because whenever we use `ml-toolbar-menu`, we have
to get a reference to a user object anyways and pass it to it as

```
<ml-toolbar-menu [user]="user"></ml-toolbar-menu>
```

This became apparent when I tried to logout on the user profile page.
While the logout was successful, `ml-toolbar-menu` didn't get notified
about the user state change because we weren't observing user changes
in `UserProfileComponent`.

Since we don't want to reimplement this observation everytime we use this
component, we simply make it smarter by letting it observe those changes
itself.