feat: enforce a 5g disk quota
With this change we are enforcing a
5 GB disk quota per execution.

This works as follows:

The container is mounted with the --read-only
flag which basically means that the user
can't make any changes.

Except, we mount a temproray filesystem
at /run which is the directory where
labs get executed.

In plain english, the user is allowed
to write a maximum of 5 GB into the
current directory (/run) but isn't allowed
to write anywhere else.

Relates to #35