feat: redirect users after login when displayName is undefined
As discussed in #99, it can happen that when a user logs in via GitHub
and there's no `displayName` in GitHub, that user also won't have a proper
`displayName` in MachineLabs either.

In addition, since recent database rules changes, the user can't even be
saved when the `displayName == null` anyways.

This commit does two things:

1. It ensures that if the gh `displayName` is `null`, the `displayName` in ML becomes `Unnamed User`
2. If that's the case, the user gets redirected to her profile so she can change her `displayName` to
  a proper one

Closes #99