feat: store invocation lab with execution
Previously we only stored the lab_id at the
execution. This can become a problem when
we have to figure out what an execution did.

For instance, we may be confronted with
a crime scenario where we just know
that an IP was spamming. Without this,
we have no way to figure out who was
responsible.