feat(ValidationService): introduce concept of Resolver
With Resolvers it is possible to decouple the
concept of validation and resolving data.
Often multiple independent validations
need to act on the same data.

Without resolvers the data needs to be
fetched from the db again and again
for every validator. With resolvers, the
data can be fetched once and is then reused
by various validators.