feat: re-introduce edit lab dialog
This is being re-added because we decided that most of the crucial configurations
of a lab will happen in a lab's `ml.yaml` introduced in https://github.com/machinelabs/machinelabs-client/commit/a3f771c96c629fe0ff2fb65526e8dbc26ceef8c5,
while meta information about the lab (name, description, tags) will
be edited in a separate view (the dialog).

The actual settings tab will be changed so that it's read only and renders
information about the settings from the `ml.yaml`.