fix: trace down execution for stop invocations
Previous to this change if a client sent an
invocation to stop an execution we took for
granted that the execution will have the id
of the invocation. However, for invocations
that get redirected to another execution this
is actually not the case.

With this change, the server traces down the
invocation until it finds the execution that
needs to be stopped.

Fixes #84