feat: redesign stop feature
With this change the stop button
moves from the toolbar to the execution
card. This not only moves the feature
visually it also changes a lot about
how stopping works.

The new stop API works on actual
executions.

As a sideffect of this, the notification
that informs about an unresponsive server
also does not offer the option to stop
anymore.

Relates to #198, #193