fix: make URL updates without navigation more robust
The function to update the editors url without actually performing a navigation
introduced in https://github.com/machinelabs/machinelabs-client/commit/1b2cec44aea3c569077298c7fca74365b80b7fe5
is going to be refactored to much more robust `UrlSerializer` APIs provided
by Angular.

We're now using exactly the same mechanisms that the router uses to update
the url just that we don't schedule a navigation.

I couldn't reproduce #204, however I think this should now be fixed with this
implementation.

Notice that this commit also fixes a bug that whenever a user selected a file in
the file browser, the url was update with the `file` query parameter but it also
erased all other existing ones. This can be a problem when we start implementing
query parameters to select executions of a lab.

Fixes #204