feat: add basic service worker support
This commit adds Angular's Service Worker capabilities to MachineLabs when building
the app for production. Here's how it goes:

- build app for production using `ng build --prod`
- inspect `/dist` folder a notive `ngsw.manifest.json`
- `index.html` has a couple new script added to it now

Basically what happens is that CLI automatically adds a manifest file for us which
is going to be consumed by the service worker, which is also automatically added
to the production builder (in addition to a script that registers the SW).

If you serve the generated dist check the servie worker panel in the devtools,
you'll see that a SW has been installed. Reload the app and check the network
panel. Static assets are now served from cache.