fix: show stop and restart button only if user owns execution
Actually we want to check if the user owns the lab. However, that
would've required a bigger refactoring as the user id isn't stored
on a lab that's attached to an execution.

That's why we simply check if the user owns the execution. This should
be enough because only owners of a lab can execute it.

Closes #216