feat(rleService): handle large outputs more efficiently
With this change the rleService optimizes better
for large outputs.

In particular, if the message count is above
a specified treshold, we only fetch head, tail
and live messages (everything coming after the tail)
and put a message in between that tells how many
messages were skipped.

This dramatically improves both loading speed
and stability of the UI.

Fixes #220