feat(LocationHelper): introduce LocationHelper service
`LocationHelper` provides helper APIs related to an application's
location. One use case is to have an Angular router like API to
update the url without triggering a full router navigation cycle.

We have this scenario when we execute an unsaved lab for the very
first time. After it gets save and before it gets executed, we want
to append the new lab id in the url.