feat: introduce HasExecutionGuard
This guard ensures that the latest execution of a lab is selected by
default, by adding the `executionId` route parameter to the url, in case
it's missing.

If a user enters a lab that doesn't have any executions, it's perfectly fine.
If a user enters a lab that has executions but non is specified in the url,
ML will take care of that.

That's why this guard will always resolve if there's

a) an execution id present in the url
b) no execution available for the selected lab

Otherwise it's going to cancel the activation and redirect to the same lab
but with the latest available execution.

Closes #241