fix: don't listen to latest execution in list when initializing lab
A regression was introduced in https://github.com/machinelabs/machinelabs-client/commit/48e2d31ce8fad8f738d7e6552544460a835ecb07
that `rleService.listen()` was executed twice in lab initialization.
First with the execution id provided via URL (the one determined by
`HasValidExecutionGuard`), and then a second time in `initLab()` where
we, up until now, simply picked the first execution of the list and
listened to it as well, which essentially caused the output to be
overwritten while the correct execution is selected.