refactor: consume execution manually without relying on route change
This commit goes back to using manual URL updates without triggering route
changes and therefore also doesn't rely on route parameter changes anymore.

One of the reasons we're doing this is because the previous architecture
introduced ugly side effects like flickering in the UI when running a lab
for the first time, due to route changes.

There's also a change in `EditorViewComponent`'s `listen()` API. Instead of
taking an entire `Execution` it only takes an execution id, as it makes the
method more reusable throughout the code base.