refactor(admin-cli): move template out of target
Semantically it doesn't make sense to have
`template` as part of target because the
template should control many configs
beyond the target one.

Unfortunately it doesn't seem to be easily
possible to get what we want with a standalone
template config so this commit moves everything
under a common cfg.* namespace.

We end up with cfg.template, cfg.env, cfg.target.* etc.