fix(server/messagingService): make sure msg index is correct
This is fixing a terrible bug that cause the
message index to keep counting up between different
executions.

What this means is that we have executions (in the live system)
where the message index starts at a really really high number.
This affects our MessageStreamOptimizer and can
cause all kind of subtle bugs.

Fixes #362