perf(client): make UndoManager configureable
The UndoManager causes a major memory leak
when performing multiple executions in a
lab. We definitely don't need it for the
output tab and most likely it's not super
important for the editor tab either.

This commit disables it by default
but one can always enable it through
the `undo` property.

Relates to #307