fix(admin-cli/cut): ensure correct yarn.lock
Cutting a release includes increasing all
version numbers including those of
shared libraries.

This in turn will cause the yarn.lock
to get updated of those projects that
use the shared libs.

This commit ensures that the updated
yarn.lock of the server will be part
of the release commit.