feat(server): continuously recycle space for infinite messages
With this commit we change the way how we limit
the number of messages per execution in our db.

Previously, we had a hard cap of 100.000 messages
per execution. If an execution produces more messages
it kept going but it's output was no further preserved.

This is really annoying because often in Machine Learning
the training will go on for hours or days producing
a really large amount of output.

Our long term strategy is to provide a way for the user
to download the entire output as zip files but we still
need to have something in the meantime.

This patch does the following:

*Birds view*

It introduces a RecycleService that basically keeps
track of all produced messages and triggers a
recycle phase every time a certain treshold of messages
was reached.

This works pretty compareable to a regular garbage
collection in the sense of the so called "stop-the-world"
strategy. This means, that during a recycle phase all
further produced messages queue up and have to wait
until the recycling phase is done before they can
continue to be processed. Recycling and writing into
the database can't happen in parallel.

There are certain tradeoffs to be made in terms of
how often recycling should happen vs how much memory
we want to allow for recycling etc. pp.

*Limitations*

The recycle phase is heavily optimized to deal with
certain limitations. For instance:

1. There is no built in way to limit `child_added`
   events so that one knows when a certain set of
   data was fully loaded.

2. There is no fast way to know the number of nodes
   that were returned. Calling `numChildren()` on
   a snapshot will also cause looping through the
   entire dataset.

For 1.) we aren't using `child_added` but `value`
instead which comes at the cost of a higher memory
and IO footprint. This means that we have to limit
our size of rows that are affected by a recycle phase.

For 2.) we avoid upfront couting and instead count
while we process the data. If after counting we
notice that something is not as expected, we skip
the recycle phase. This shouldn't happen in practice
but it's part of a defensive programming strategy.

*The recycle phase in detail*

When a recycle phase is triggered (currently when
7000 messages are reached) we do the following.

1. We read all messages from the tail of the data.
   This is currently set to read the last 6000
   messages.

2. We generate one atomic bulk command that does

A.) delete the first 5000 messages
B.) patches the index of the last 1000 messages

We run this command only if the count matches
our assumptions.

3. When the command was applied we reset the
   `index` for all further messages that are
   written into the db up from that point on.

Notice that we keep track of a `virtual_index`
that each message gets assigned so that one
knows where in the entire log of the process
that message appeared. It's also helpful
to know how many messages where deleted.

Also notice that we still keep the hard limit
as kind of a "last resort" should anything
fail with this new approach. However, we
should reduce the hard limit to something
like 10.000 messages with another commit.

Fixes #395