fix: fix production build by moving autotrack to vendor scripts
Turns out autotrack was already minified which causes our production build
to fail. This commit removes imports to autotrack and adds autotrack as a
vendor script to CLIs config. That's where it should be anyways.

Also it turned out autotrack already comes with all plugins included, there's
not need to import/embedded urlChangeTracker explicitely.

In the future we might want to revisit this and have a custom build for autotrack
that doesn't always include all plugins (which is the case right now).

However filesize is still kind of low (8kb gzipped) so I'm going to
postpone that for now.