fix(server): ensure to kill execution on unexpected errors
With this patch we make sure to kill the execution
in case if an error bubbled all the way up to kill
the message stream.

This case should actually not happen but in case
it happens due to some bug in the system, we need
to make sure to kill the execution and set a failed
state.