fix(server): don't adjust index for last message
We previously adjusted the index in cases where
we had reached the maximum capacity of our
messages node so that when the process kept
going and we finally reached the last message
we would reflect its proper index. However
now that we recycle messages the index should
always reflect the index in the database.

We have the `virtual_index` to reflect the
index in the entire space of messages.