chore: add migration script
This script migrates existing labs to align with the latest schema changes:

- `Lab` now has a `hidden` property (default `false`)
- New `user_visible_labs` index represents all visible labs for a user

Prior to this commit we don't didn't have a `hidden` property in labs.
That means the decent default for all labs that need migration is adding
a `hidden` property to all existing labs and set it to `false`.

At the same time, all labs that will be touched can also be added to the
new `user_visible_labs` index, as all labs will be visible by default.

To run this script execute the following command from `./admin-cli`

```
$ npm run mla -- migrate --file=./firebase/migrations/397/set-labs-hidden.js --cfg.template=SOME_TEMPLATE
```

Ideally, add the `--dry-run` option to double-check.