feat(server): refactor docker usage
This uses our new reactive process tools to
utilize the docker API in more fine grained
steps.

Instead of one huge "docker run" that does
it all, we know do individual commands to

1. Create a container

2. Start the container

3. Run the execution code

4. Destroy the container

In the future this will allow us to request
a new API token for the Google storage API
to upoad files right after the execution
finished.

Relates to #387