feat: make active tab configurable via URL query parameter
This commit allows users to do things like:

```
/my-lab?tab=editor
```
or
```
/my-lab?tab=console
```

Also, the `tab` query parameter updates automatically when clicking on
tabs.

*ATTENTION*

This also removes the behaviour ob labs automatically selecting the console
tab when initializing an execution. This only happens if no `tab` queryParam
is given, enabling us to stay in the editor tab.