feat(client): reword rejection dialog
This change tunes the wording of the rejection
dialog for anonymous users a bit.

Users are more likely to accept a rejection
if they know the reason for it (studies prove that).

So with this rewording, we are telling the user
that we require the login to prevent abuse
of the platform.