feat(supervisor): implement lib with super powers
This libray implements services that have special
privileges to control the system.

This implements a first `TakedownService`
with power to take down executions by the system.

At first, this will be used in the admin-cli.
Later on, this will also be used in a cron job
that permanently monitors the executions to
ensure executions are stopped by the system
when the user runs out of free credits.

Relates to #366