fix: ensure snack messages work when output isn't consumed
With this change, the "controlling messages" such as
ExecutionRejected, ExecutionStarted and ExecutionFinished
are consumed separately from the STDOUT/STDERR messages
so that the editorService can still raise snackbar
messages even if the actual output isn't consumed
just yet.

This also moves the LateExecutionMessage into the
editorService to ensure the snackbar still works
which comes with the added benefit of being
able to show this message even when in pause mode.

Relates to #493