feat(server): implement mount feature
With this change files from our bucket can
be mounted at runtime. The syntax is as follows.

mounts:

mounts:
    - path: <user>/<mount>/<version>
      name: <name-of-local-directory>

As an example:

mounts:
     - path: machinelabs/mnist/1
       name: my-data

Notice that the path may start with a leading
slash which is ignored. Notice also that the
version can be omitted which will result in
defaulting to the latest version.

The name is not allowed to contain any path
delimiter. It must be a legal directory name.

Relates to #564, #488