chore(client): distribute shared libs in JS via dist
Previously our shared libs did not have a dedicated
dist folder where the compiled JS would go.

This setup introduces problems down the road and
is discouraged.

See:
https://github.com/angular/angular-cli/issues/8284#issuecomment-341417325

This change migrates all shared libs to the new
layout.

This change has several implications:

1. There is a new special `build.js` in `./shared` that builds
all shared libs in the right order. There's an alias admin-cli
taks for this but the script itself needs to live outside the
admin-cli because the admin-cli itself depends on the shared libs,
hence *building* the shared libs can not depend on the admin-cli.

2. Building the shared libs is currently quite costly. Triggering
implicitly was slowing down the build and general feedback loop
for the development. Therefore the shared libs now need to be build
explicitly. This can be done either by running `build.js` in `./shared`
or running the `build-shared` command of the admin-cli.

Notice that building the shared libs does not automatically update
the libs in client and server. For the client there is an npm script
called `updated-shared` that can be run to update the libs. However, the
server does trigger a build for the shared libs and also updated the
packages. However, since this slows down the build quite a bit this
can be skipped by passing --skip-shared` when building the server.