fix(toObservableProcess): handle and propagate spawn errors
Previously we only propagates errors from STDERR but
not errors that come from the spawning itself. For
instance if you spawn a command `foo` where `foo`
does not exist, the spawning will fail but it's not
`foo` that is yielding STDERR messages but the
spawning itself which errored because `foo`
does not exist.

This change handles and propages such cases so that
any error will result in a message of OutputType.Sterr
followed by closing the stream.