refactor(ExecutionRejectionReason): use string enum
Migrating to an string enum may look like a huge
breaking change (in fact, it is) but in reality
it shouldn't cause big trouble.

While it is true that all existing rejected
executions will can't be matched anymore,
those messages are throw away one-time information
only visible to the one who started the exectution
and only in the very moment of the rejection.

And even if for some reason client and server
talk different things for a brief moment in time
the worst thing that could happen is the client
showing a generic rejection message instead
of a specialized one.

The benefits outweigh the risks.