fix(server): ensure server shows correct execution state
We previously did not show the correct execution
state when an execution was stopped. This change
ensures that stopped executions will have the
state set to `stopped` correctly. More importantly
though, it also ensures that all states except
`executing` are treated as final. This means,
once an execution is in any other state than
`executing`, the state can not be changed anymore.
Ignoring `Pristine` here as that is an artificial
client side state that doesn't have any meaning
in the db or on the server.

This is important because our current system
could otherwise lead to bugs where we set an
execution to stop, just to have it overwritten
briefly after when the process actually completes.

Fixes #259