fix(EditorService): ensure activeFilePath is set when opening file
As part of https://github.com/machinelabs/machinelabs/commit/5db9b0537324a02c50d3f944b923b45d7c3e7359 we've introduced a regression that sets
an invalid state in our `EditorService`, where the `activeFilePath`
is not set on lab initalization, when no `file` query parameter is
given.

This fix ensures that `activeFilePath` always gets a value when a
file is opened - either the full path of the file to be opened, or,
if a lab is opened without a file path, just the file name itself.
This usually means the default entry file (right now `main.py`) is set.

For more information see:

https://github.com/machinelabs/machinelabs/issues/570#issuecomment-346023151

Fixes #570