fix(LabExecutionService): ensure we have no duplicates in execution list updates
Some changes on execution objects (like setting `hidden` state)
cause changes in our database indices that trigger `child_added` events
and make our execution list re-add the same execution object multiple
times.

This caused a bug when executions have been removed through the UI and
then "undo-ed", that the same execution showed up 2 - n times.

This commit ensures that, when an execution is being pushed in the execution
list, we first check if that same execution already exists in the list. If
that's the case, we remove the first item again.

For more information:

https://github.com/machinelabs/machinelabs/issues/632#issuecomment-347498709

Fixes #632