fix: set contentType on uploads to empty string
The contentType used to exist on the bucket event
object but doesn't seem to exist anymore. This
never really worked as the contentType always
used to be an empty string but now that it is
undefined it's crashing our cloud function.

This change sets it to empty string until we
fix that properly. The reason we prefer an
empty string over removing it entirely, is, that
now all depending code can safely assume there
is at least a field content_type of type string.