fix(EditorView): ensure MonacoEditor's models are disposed before initializing lab
This commit fixes a regression introduced in https://github.com/machinelabs/machinelabs/commit/e7c646d7f7b984e83c60024d376681f308ca3cfe where newly created labs don't render
the correct file contents because the `MonacoEditor` keeps references to file contents
by name in its own cache.

This means that, every time `MonacoEditor` opens a file (which is when a user selects
a file, or MachineLabs changes the lab model and therefore has a new active file),
it retrieves the contents of that file from its own model cache by the name of that file.

That caused problems when a user created a new lab from within the editor view, because
all MachineLabs does is updating the lab model and binding a new active file to the
editor (most of the time `main.py`), because the file contents of the previous lab where
shown.

This commit ensures that, whenever a new lab is created, that all `MonacoEditor` models
are disposed first. This way we keep the cache for when users stay in a single lab, but
we free it up when the whole lab is a different one.

Fixes #667