feat(cli): amend lab if it exists
With this change, the CLI will attempt
to amend a lab if it already exists.

In practice, this means that one does
not have to care about `name`, `description`,
`tags` etc as long as one does not want
to change them. They will be taken
from the existing lab.

Only if one wants to overwrite `name`
or `description` one can specify
`--name` or `--description` to set
them via flag.

Closes #677