fix(LabStorageService): ensure to emit empty list if no recent labs are available
This commit fixes a regression that if there are no recent labs in
`/idx/recent_labs`, `getRecentLabs()` will never emit a value, due
to `forkJoin()` getting an empty array of observables.

This is not a problem in production because we will probably never
erase the index, however it's still a bug.