fix(server): ensure executions can't memory attack server
Without this our system is currently vulnerable against
memory attacks.

Attacks are quite simple as we allow executions to
consume as much memory as they want. At some point
the execution consumes as much memory that the host
starts shutting down processes including our
server process or just any other random process.

This fix mitigates the risk by introducing an
upper bound for the overall kernel memory
that can be consumed.

Basically what this means is that each container
can still consume as much memory as it wants
but there's a threshold of what is reserved
as kernel memory.

https://docs.docker.com/engine/admin/resource_constraints/#--kernel-memory-details

Fixes #687