fix: make lab template object immutable
This commit fixes a bug in our app that, whenever we change the file contents
of a lab that has been created by a template, we mutated the template instead
of working on a deep copy.

We now ensure that we deep clone lab templates so we don't mutate the original
ones.