feat(server): introduce concept of transformers
Transformers run right after values have been
resolved and have access to all resolved
values that exist at that time. This allows
to transform resolved values before certain
rules work with them. Notice that each
transformer runs on every resolve and
that our validation system does not guarantee
that all resolves run before all rules.

This is a feature that ensures we do not
resolve unnecessary because it allows
rules to fail early and cut of the resolve
pipeline. In other words resolves and
rules may run at the same time and the
only guarantees that our system gives are:

- Rules can depend on resolves
- The first rule that needs a resolver will
  implicitly trigger it
- Resolved values are always shared and never
  resolved twice