chore(core): get rid of shell-based lab persistence APIs
These APIs were based on the idea that we could generate
shell commands to write labs to a filesystem. This was
a convenient way to write files from a host into a
docker container without mapping actual volumes.

However, we later became aware that shell commands are
limited in maximum size:

https://github.com/machinelabs/machinelabs/issues/713

Hence, the shell based lab persistance APIs don't
seem to have any place for usage anymore.

Relates to #713