feat(admin-cli): introduce new generate-changelog command
THe command can be either used as standalone command or as part of
the admin CLI's `cut` command.

Example:

```
$ mla generate-changelog --version foo [--dry-run] [--release-count=1]
```

The `--dry-run` option will output the generated changelog to stdout.
Otherwise the latest release will be prepended to `CHANGELOG.md`.
If `CHANGELOG.md` doesn't exist, it'll be created.

The `--release-count` option can be used to control for how many past
releases a changelog should be generated. Default is 1, meaning basically
all commits from last git tag til `HEAD`. When set to 0, a complete
changelog is generated. This is useful when generating a changelog for
the first time.