fix(firebase/functions): fix deploy script
We introduced a regression in https://github.com/machinelabs/machinelabs/commit/f7d15eba3733fbe0424b32e894cc3f787b5fdc39 that,
when `yarn` isn't globally installed anymore the firebase functions' deploy
script will throw an error. This is not a problem when `yarn` is globally
available, which should be the case for most of us as of now.

However, since we switched to `npx` it's probably better to streamline
the deploy script as well and never assume a global `yarn` is available.