fix(client): ensure embedded editor view fills available viewport
As discussed in #792, we've introduced a bug in https://github.com/machinelabs/machinelabs/commit/f0daa615ea54d7726023d7e75213c54eba67c8a3, that the editor panels won't fill the available viewport anymore in embedded editor
view.

This is because we've removed `flex: 1`, so the main content doesn't
scroll horizontally when the file drawer is opened.

This commit ensures the editor layout fills the available space again.

Fixes #792