chore(client): remove share dialog
While it's nice to have this, all it really does is opening a dialog
that displays the current URL so people can copy it to share it somewhere.

I don't have any scientific data to back this, but I assume people don't
really use this button. In fact, I don't see anyone really sharing labs.

By removing this button, we have room for a more important UI feature,
which is a button that gives users the opportunity to leave feedback.