chore(CI): enforce yarn version 1.6.0
This is considered best practice to ensure builds are working across the
board. In fact, right now we have the problem that there's a bug in 1.7.0
which prevents our builds to pass. Until 1.8.0 is released, we have to nail
the version to 1.6.0 as CI uses the latest available version by default.